Extend all size measurements beyond MRCA
So far, the size measurements stopped at the most recent common ancestor
(MRCA). This meant that as we went back in time, the number of samples
from which the quantiles were computed went down to eventually 1. These
last samples invariably had large population sizes, because the time to
the most recent common ancestor is proportional to the population
size. In other words, the estimates became biased upwards as we moved
into the past. This is now fixed by pretending that all measurements
extend indefinitely beyond the most recent common ancestor.