Changed plotting to step-wise
In the past, epos -r and epos would give different plots when applied to
a single epos result. Output with the -r (raw) option would result in a
stepped plot, output without -r in a continuous plot. Now the results
are identical in both cases. Stepping required changes in the
print.PrintQuantiles function and adjustment of the documentation. In
addition, the epos output has changed in the meantime and hence a new
example data set was generated and the documentation adjusted accordingly.